refactor: move reader table generation to module in doc/source
