+++ .gitignore
@@ -74,3 +74,4 @@ doc/source/_build/*
 # this should be generated automatically when installed
 satpy/version.py
 doc/source/api/*.rst
+doc/source/reader_table.rst
