+++ satpy/utils.py
@@ -33,7 +33,6 @@ import yaml
 from yaml import BaseLoader
 
 from satpy import CHUNK_SIZE
-from satpy.readers import available_readers
 
 try:
     from yaml import UnsafeLoader
@@ -623,63 +622,3 @@ def _merge_storage_options(storage_options, storage_opt_dict):
         storage_options = storage_opt_dict
 
     return storage_options
-
-
-def rst_table_row(columns=None):
-    """Create one row for a rst table.
-
-    Args:
-        columns (list[str]): Content of each column.
-
-    Returns:
-        str
-    """
-    row = "    * - {}\n".format(columns[0])
-    columns = ["      - {}\n".format(col) for col in columns[1:]]
-    row = row + "".join(columns)
-
-    return row
-
-
-def rst_table_header(name=None, header=None, header_rows=1, widths="auto"):
-    """Create header for rst table.
-
-    Args:
-        name (str): Name of the table
-        header (list[str]): Column names
-        header-rows (int): Number of header rows
-        width (optional[list[int]]): Width of each column as a list. If not specified
-            defaults to auto and will therefore determined by the backend
-            (see <https://docutils.sourceforge.io/docs/ref/rst/directives.html#table>)
-
-    Returns:
-        str
-    """
-    if isinstance(widths, list):
-        widths = " ".join([str(w) for w in widths])
-
-    header = rst_table_row(header)
-
-    table_header = (f".. list-table:: {name}\n"
-                    f"    :header-rows: {header_rows}\n"
-                    f"    :widths: {widths}\n\n"
-                    f"{header}")
-
-    return table_header
-
-
-def reader_table():
-    """Create reader table from reader yaml config files.
-
-    Returns:
-        str
-    """
-    table = [rst_table_header("Satpy Readers", header=["Description", "Reader name", "Status", "fsspec support"],
-                              widths=[45, 25, 30, 30])]
-
-    reader_configs = available_readers(as_dict=True)
-    for rc in reader_configs:
-        table.append(rst_table_row([rc.get("long_name", "").rstrip("\n"), rc.get("name", ""),
-                                    rc.get("status", ""), rc.get("supports_fsspec", "false")]))
-
-    return "".join(table)
